Typing fix

+++ src/cbor_internal.h
@@ -72,7 +72,7 @@ void _cbor_handle_load_uint64(const unsigned char * source,
 
 void _cbor_handle_load_bytestring(const unsigned char * source,
 								  size_t source_size,
-								  uint64_t length,
+								  size_t length,
 								  cbor_item_t * item,
 								  struct cbor_load_result * result);
 #endif
